[RENAME] : ✏️: yourmediaapp에서 tviewing으로 전환
