+++ app/src/androidTest/java/com/example/tviewing/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.yourmediaapp_18
+package com.example.tviewing
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.yourmediaapp_18", appContext.packageName)
+        assertEquals("com.example.tviewing", appContext.packageName)
     }
 }
