+++ app/src/main/res/values/themes.xml
@@ -1,9 +1,9 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.YourMediaApp_18" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.TViewing" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.YourMediaApp_18" parent="Base.Theme.YourMediaApp_18" />
+    <style name="Theme.TViewing" parent="Base.Theme.TViewing" />
 </resources>
