+++ app/src/test/java/com/example/tviewing/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.yourmediaapp_18
+package com.example.tviewing
 
 import org.junit.Test
